statusbar auto, fix cookie policy image
